+++ tapo/src/lib.rs
@@ -17,7 +17,7 @@
 //!     let tapo_password = env::var("TAPO_PASSWORD")?;
 //!     let ip_address = env::var("IP_ADDRESS")?;
 //!
-//!     let device = ApiClient::new(tapo_username, tapo_password)?
+//!     let device = ApiClient::new(tapo_username, tapo_password)
 //!         .l530(ip_address)
 //!         .await?;
 //!